Add postcode and pupils to Branch response model
This brings it in line with the current Branch API interface where we've
recently added the postcode property.

In theory, all properties should be considered optional as they can be
defined with request filters (in exactly the same way as it's done in
GraphQL) but for now only the relational properties are marked as such
for simplicity's sake.

Also, the model has been renamed to emphasise the model it represents
rather than the context which it is used in (ie.: `PupilsByDayByBranch`).