Add request for attendances for given branch and date
This is required to fetch the list all of clocked on pupils, including
those that are no longer active and therefore would not be included in
the daily register.

The retrieval of the daily register will now be a multi-step process
because we don't have a concept of the active/inactive pupil timeline;
we only have a static active Y/N property on the pupil object.