Add `buildParentRegisterRequestFilter` helper
We were having issues with conformance to the Decodable protocol in
one of the nested structures in `ParentRegisterRequestFilter` so we
tried nesting them and that, for some reason, resolved them. So we
followed that pattern in other filter models: all structs are now nested
inside of those representing the `scope` noded.