Merge Attendances and Parents relations for reuse
Treating these types as separate proved to be challenging when trying to
encode arrays containing both of them. The solution in the previous
commit meant you would not be able to define an array with multiple
types and although this wouldn't be an immediate problem at this moment
in time, it could prove to be problematic in the future. This problem
has been eliminated with this change.