Make JSON unit tests less strict
In previous versions of Xcode/Swift creating JSON would always yield the
same string. But this doesn't seem to be the case any more so comparing
against a pre-stringified value started causing tests to fail.

For that reason we've made some tests a bit looser - although they are
not checking the exact query strings, they do confirm if some expected
values are present in the URL.